+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Battleship".
 - Tests for "Texas holdem".
 - Tests for "Solid integer".
+- Tests for "Bit count to limit".
 
 ## [1.15.0] - 2022-11-30
 ### Added